requirements.txt
asgiref==3.2.5
btrdb==5.6.1
conduit==1.1
Django==3.0.4
django-crispy-forms==1.9.0
django-tempus-dominus==5.1.2.13
grpcio==1.16.1
grpcio-tools==1.16.1
joblib==0.14.1
mysql-connector==2.2.9
mysqlclient==1.4.6
numpy==1.18.2
pandas==1.0.3
Pillow==7.0.0
protobuf==3.11.3
PyMySQL==0.9.3
python-dateutil==2.8.1
pytz==2019.3
PyYAML==5.3.1
scipy==1.4.1
six==1.14.0
sqlparse==0.3.1
tabulate==0.8.7
